+++ src/components/Home.js
@@ -1,12 +1,8 @@
 
 import React, { useEffect, useState, lazy, Suspense } from 'react'
-import styled from 'styled-components'
-import { throttle } from 'underscore'
 import { Helmet } from 'react-helmet'
 
-import { Typography, Link, InputBase, Paper, IconButton, Avatar } from '@material-ui/core'
-import SearchIcon from '@material-ui/icons/Search';
-import Autocomplete from '@material-ui/lab/Autocomplete'
+import { Typography, Link } from '@material-ui/core'
 import { MainContentContainer } from './MainContent'
 import Breadcrumb from './Breadcrumb'
 import { handleFetch } from '../helper/FetchHandler'
@@ -16,26 +12,10 @@ import OneThirdTwoThirdsGrid from './util/grid/OneThirdTwoThirdsGrid'
 
 import { LeftBoxPaper, RightBoxPaper, RightBoxHeaderTypography, RightBoxSubHeaderTypography } from './util/grid/OneThirdTwoThirdsGrid'
 
-import {RenderGroup, SearchSuggestionTypography} from './util/Search'
-
-import Parent from './Parent'
 import {HEART_API_HOST_NAME} from '../helper/DownstreamServices'
+
 const YouTubeUploads = lazy(() => import('./YouTubeUploads'))
-
-const DatabaseSearch = styled(Autocomplete)`
-	&&&
-	{
-		.MuiAutocomplete-popper
-		{
-			background-color: black;
-		}
-	}
-`
-
-
-const searchThrottle = throttle((searchSubject, setSearchOptions, history) => {
-	handleFetch(`${NAME_maps_ENDPOINT['search']}?limit=10&cName=${searchSubject}`, history, json => { setSearchOptions(json) })
-}, 15)
+const DatabaseSearch = lazy(() => import('./DatabaseSearch'))
 
 
 export default function Home({ history }) {
@@ -43,10 +23,6 @@ export default function Home({ history }) {
 	const [banListTotal, setBanListTotal] = useState(0)
 	const [yearsOfBanListCoverage, setYearsOfBanListCoverage] = useState(0)
 	const [productTotal, setProductTotal] = useState(0)
-
-
-	const [searchInput, setSearchInput] = useState('')
-	const [searchOptions, setSearchOptions] = useState([])
 
 	const [youtubeData, setYoutubeData] = useState(undefined)
 
@@ -65,15 +41,6 @@ export default function Home({ history }) {
 	}, [history])
 
 
-
-	useEffect(() => {
-		if (searchInput !== '') {
-			searchThrottle(searchInput, setSearchOptions, history)
-		}
-	}
-		, [searchInput, history])
-
-
 	return (
 		<MainContentContainer>
 			<Helmet>
@@ -87,69 +54,8 @@ export default function Home({ history }) {
 
 			<Breadcrumb crumbs={['Home']} />
 
-
-			<Paper style={{ display: 'flex', width: '95%', maxWidth: '500px', margin: '0 auto', borderRadius: '1.5rem', marginBottom: '2rem' }} >
-				<DatabaseSearch
-					id='search'
-					clearOnEscape
-					selectOnFocus
-					style={{ flex: '1' }}
-					noOptionsText={(searchInput === '') ? 'Type For Suggestions' : 'No Results'}
-					getOptionLabel={option => option.cardName}
-					options={searchOptions}
-					groupBy={option => option.cardColor}
-					onChange={ (event, value, reason) => {
-						if (reason === 'select-option') { window.location.assign(`/card/${value.cardID}`) }
-					}
-					}
-					renderGroup={option => {
-						return (
-							<RenderGroup
-								group={option.group}
-								children={option.children}
-							/>
-						)
-					}}
-					renderInput={(params) => (
-						<div style={{ width: '100%', display: 'flex', padding: '.25rem' }} >
-							<InputBase
-								ref={params.InputProps.ref}
-								inputProps={params.inputProps}
-								style={{ color: 'rgba(0,0,0,.90)', flex: '1', margin: '.65rem', fontSize: '1.1rem' }}
-								placeholder='Search For Cards In Database...'
-								onChange={event => { setSearchInput(event.target.value) }}
-							/>
-							<IconButton >
-								<SearchIcon />
-							</IconButton>
-						</div>
-					)}
-					renderOption={option => {
-						const CARD_NAME = option.cardName
-						const UPPERCASE_CARD_NAME = CARD_NAME.toUpperCase()
-						const UPPERCASE_SEARCH_TERM = searchInput.toUpperCase()
-
-						const INDEX_OF_SEARCH_TERM = UPPERCASE_CARD_NAME.indexOf(UPPERCASE_SEARCH_TERM)
-						const LENGTH_OF_SEARCH_TERM = UPPERCASE_SEARCH_TERM.length
-
-						return (
-							<div style={{display: 'flex'}} >
-								<Avatar alt={`${CARD_NAME}-Avatar`} src={`https://images.thesupremekingscastle.com/${option.cardID}.jpg`} style={{marginRight: '.5rem'}} />
-								<div style={{ padding: '0rem', margin: '0rem' }} >
-									<SearchSuggestionTypography variant='body1'>
-										{CARD_NAME.slice(0, INDEX_OF_SEARCH_TERM)}
-										<strong>{CARD_NAME.slice(INDEX_OF_SEARCH_TERM, INDEX_OF_SEARCH_TERM + LENGTH_OF_SEARCH_TERM)}</strong>
-										{CARD_NAME.slice(INDEX_OF_SEARCH_TERM + LENGTH_OF_SEARCH_TERM)}
-									</SearchSuggestionTypography>
-									<SearchSuggestionTypography variant='body1' style={{ color: 'rgb(81,99,114)' }} >{option.monsterType}</SearchSuggestionTypography>
-								</div>
-							</div>
-						)
-					}}
-				/>
-			</Paper>
-
 			<Suspense>
+				<DatabaseSearch history={history} />
 				<YouTubeUploads history={history} youtubeData={youtubeData} />
 			</Suspense>
